ref: move api docs out of completion into the documentation module
PhelApiDocumentation and PhelBasicDocumentation lived under completion/ but had
no consumer inside it — only annotator/ and documentation/ used them, and neither
file depends on completion at all. Same reasoning that keeps the registry out of
completion: shared code does not belong in a feature package that never uses it.

Move both beside their real consumer in documentation/resolvers/, which empties
completion/documentation/ entirely.

The annotator only ever called PhelApiDocumentation.hasDocumentation(), which is
PhelFunctionRegistry.getFunction(name) != null — a registry question routed
through a documentation object. It now asks the registry directly, so the edge
disappears rather than following the move. That left hasDocumentation() unused;
drop it along with a comment justifying lazy rendering by the highlighting path
it no longer serves.
